Moved some endpoints to the right routes
